requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
dataclass-wizard==0.34.0
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycryptodome==3.21.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3